Mummi and bilbo contribution!! veriverse contribution!! LCTV contribution!! catherinacortega, sjpablo (Pawel Kulesza), fuel (Benjamin Kuhl), r00tkiller (Alexandre),  vapaamies (Vladislav Javadov), epryka (Vova - Vladimir), theaifam5 (Mateus), doraemon (currently... also gosu - hiiamgosu, dum - noticemesenpaiiamdum..., litAsHellhahah), mushmouth (Patrick McFarling),  sonnystar (sonnystaminecraft, Vlad Ionescu), abbyyyyy, alexcode, anita, ayvatan, b3n0it, bigboom, blackhatisme, boza, camwhite (Cameron White - YOSHICREW | ng2/node + mixing fire, Yoshicrew.com-out-loud contribution...), chrifrachid, codeotea, cookiesliyr, chubz, cubor, danielv, david_joffe, denvercowboy, diagram, dopedopedope, dzetko, dzhekson6000, euphemisminc, fauker, ffabiorj, georgeb, gillie, gzj71ck4elm2,  hudwan, intelijens (Michael Bui), ishaan007lucky, jarzi, jilizart, johnyblaze, jookeer, kaeris, karrmarr, kaznowski, kissloryshy, kranalex, LaffayetteBrittoPhoto (streamer with its C300 photo), magrib, marianyc, matroskinb, michaeljcalkins, michaelphp, monicaa,  nuke, paezrice, peter_cados, phanxgames (Henry Price), purpleizlegitx, rooney2, sanyakozmey, silvaesouza, siox, skyfoxx, smart6307 (thunder will/canal2048), sova, systemis, tarikogut, thiagovgamg, trello, usandy, v0n3x, wanket, wapaca, web2ls, yanaseattle, zaaz and zzzel (Marko Zel)  contribution!! jiangyubo530, kozmo, contribution!! Twitch contribution!! AlfredoPlays (and his mates: Macky43 ... ), Tectuma (GothicatCat1(Stacie), Tectuma(Kevin), Matty and Dylan), Above_Perfection, Bialyorzeu, cary205, cpmgamingltd, esmc103, GameDevCompany (Björn), GrubbyJuicde, Handmadehero (Casey Muratori), Letorusii, Mattyplays6, michaeljcalkins, nothings2 (Sean Barrett), Shathan, Snipe0803 and xandergamingtwich contribution!! Beam.pro contribution!! RoxenRox and Xellar contribution!! bisba contribution!! caro  and lucinho contribution!!
Thank you Camilo!

bilbo at veriverse
i7 6700 32GB RAM Windows 10 Pro
(booting also macOS Sierra, Hyper-V and Ubuntu 16.04 Unity Numix Theme Numix Circle Icons)

+++ appseed/axis/net/sockets/bsd/basic/sockets_tcp_socket.cpp
@@ -770,7 +770,7 @@ namespace sockets
                struct timeval tv;
                tv.tv_sec = 1;
                tv.tv_usec = 0;
-               ::select(GetSocket() + 1, NULL, &w, &e, &tv);
+               ::select((int) (GetSocket() + 1), NULL, &w, &e, &tv);
                goto retry;
             }
